Seperates templating and callback logic
