+++ egon/visualize/assets/default_style.yml
